First fumblly steps towards fixing card / unified card model names.
