+++ app/models/card.rb
@@ -1,40 +1,42 @@
 # frozen_string_literal: true
-
 class Card < ApplicationRecord
-  belongs_to :side
-  belongs_to :faction
+  self.table_name = 'unified_cards'
+  include CardAbilities
+
+  self.primary_key = :id
+
   belongs_to :card_type
-  has_many :card_card_subtypes
-  has_many :card_subtypes, :through => :card_card_subtypes
-  has_many :printings
-  has_many :unified_printings
-  has_many :card_pool_cards
-  has_many :card_pools, :through => :card_pool_cards
-  has_many :restriction_card_banned
-  has_many :restriction_card_global_penalty
-  has_many :restriction_card_points, class_name: 'RestrictionCardPoints'
-  has_many :restriction_card_restricted
-  has_many :restriction_card_universal_faction_cost
-  # This is here to support restrictions, but isn't usable on it's own.
-  has_many :unified_restrictions
-  # Will be all restrictions that directly reference this card in any way.
-  has_many :restrictions, -> { where('unified_restrictions.in_restriction': true) }, :through => :unified_restrictions
+  belongs_to :faction
+  belongs_to :side
 
-  validates :name, uniqueness: true
+  scope :by_card_cycle, ->(card_cycle_id) {
+    where(id: UnifiedPrinting.select(:card_id).where(card_cycle_id: card_cycle_id))
+  }
+  scope :by_card_set, ->(card_set_id) {
+    where(id: UnifiedPrinting.select(:card_id).where(card_set_id: card_set_id))
+  }
 
-  def advancement_requirement
-    self[:advancement_requirement] == -1 ? 'X' : self[:advancement_requirement]
-  end
-   def link_provided
-    self[:link_provided] == -1 ? 'X' : self[:link_provided]
-  end
-  def mu_provided
-    self[:mu_provided] == -1 ? 'X' : self[:mu_provided]
-  end
-  def num_printed_subroutines
-    self[:num_printed_subroutines] == -1 ? 'X' : self[:num_printed_subroutines]
-  end
-  def recurring_credits_provided
-    self[:recurring_credits_provided] == -1 ? 'X' : self[:recurring_credits_provided]
-  end
+  has_many :card_card_subtypes,
+           primary_key: :id,
+           foreign_key: :card_id
+
+  has_many :card_subtypes, through: :card_card_subtypes
+
+  has_many :unified_printings,
+           primary_key: :id,
+           foreign_key: :card_id
+
+  has_many :card_cycles, through: :unified_printings
+
+  has_many :rulings,
+           primary_key: :id,
+           foreign_key: :card_id
+
+  # Private decks
+  has_many :deck_cards, primary_key: :id, foreign_key: :card_id
+  has_many :decks, through: :deck_cards, primary_key: :id, foreign_key: :card_id
+
+  # Public decklists
+  has_many :decklist_cards, primary_key: :id, foreign_key: :card_id
+  has_many :decklists, through: :decklist_cards, primary_key: :id, foreign_key: :card_id
 end
