+++ lib/tasks/import_decklists.rake
@@ -33,7 +33,7 @@ namespace :import_decklists do
       printing_to_card[printing.id] = printing.card_id
     end
 
-    cards = Card.all
+    cards = UnifiedCard.all
     cards_by_id = {}
     cards.each do |card|
       cards_by_id[card.id] = card
