+++ test/unit/deck_validator_test.rb
@@ -1,3 +1,5 @@
+require 'test_helper'
+
 class DeckValidatorTest < ActiveSupport::TestCase
   def setup
     @empty_deck = {}
@@ -740,6 +742,7 @@ class DeckValidatorTest < ActiveSupport::TestCase
     v = DeckValidator.new(deck)
     assert !v.is_valid?
     assert_equal v.validations.size, deck['validations'].size
+    puts v.validations[0].inspect
     assert_includes v.validations[0].errors, 'Card `trieste_model_bioroids` is banned in restriction `standard_banlist`.'
   end
